fix: ATLAS-282: Separate Extension code from primary Entity class
